refactor: extracted searchbar into component
