+++ src/components/index.js
@@ -6,3 +6,4 @@ export { VideoCard } from "../components/video-card/VideoCard";
 export { PrivateRoute } from "../components/auth/PrivateRoute";
 export { RestrictedRoute } from "../components/auth/RestrictedRoute";
 export { Modal } from "../components/modal/Modal";
+export { SearchBar } from "../components/searchbar/SearchBar";
